debug

+++ src/inference/inference.cu
@@ -58,13 +58,13 @@ void printCudaMemoryInfo() {
 }
 
 // Kernel to check and print the embeddings
-__global__ void check_embedding(float *fp16_tensor, int dim) {
+__global__ void check_embedding(__half *fp16_tensor, int dim) {
     for (int token_idx = 0; token_idx < d_NUM_TOKENS; token_idx++) {
         printf("Token %d embeddings:\n", token_idx + 1);
         int max = 0;
         float curr_max = 0.0f;
         for (int i = 0; i < dim; i++) {
-            float embedding = fp16_tensor[token_idx * dim + i];
+            float embedding = __half2float(fp16_tensor[token_idx * dim + i]);
 
             if (embedding > curr_max) {
                 curr_max = embedding;
